performance increase in database removing dups

+++ build.sh
@@ -2,23 +2,27 @@
 
 # Building libtwitcurl
 echo "Building libtwitcurl"
-cd libtwitcurl
-cmake .
-make -j4
-mv libtwitcurl.a ../src/
-if [ $? -ne 0 ]
+if [ ! -f ./src/libtwitcurl.a ]
 then
-    echo "Failed while compiling libtwitcurl"
-    exit 1
+    cd libtwitcurl
+    cmake .
+    make -j4
+    mv libtwitcurl.a ../src/
+    if [ $? -ne 0 ]
+    then
+        echo "Failed while compiling libtwitcurl"
+        exit 1
+    fi
+    cd ..
 fi
 
 # Building twitterfu
 echo "Building twitterfu"
-cd ../src
+cd src/
 SOURCES=$(ls *.cpp)
 for S in $SOURCES
 do
-    g++ -Wall -c -Werror $S
+    g++ -Wall -c -Werror -O4 -std=gnu++11 $S
 done
 
 if [ $? -ne 0 ]
@@ -27,7 +31,7 @@ then
     exit 1
 fi
 
-g++ *.o libtwitcurl.a -lsqlite3 -lcurl -o twitterfu
+g++ *.o libtwitcurl.a -lsqlite3 -lcurl -O4 -o twitterfu
 
 if [ $? -ne 0 ]
 then
